requirements.txt
amqp==5.2.0
asarPy==1.0.1
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
kombu==5.3.7
Markdown==3.4.4
pika==1.3.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
